# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Duka POS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MpesaCallbackPayloadBody = { [key: string]: unknown };

export interface MpesaCallbackPayload {
  Body: MpesaCallbackPayloadBody;
  [key: string]: unknown;
}

export interface MpesaCallbackAck {
  ResultCode: number;
  ResultDesc: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  username: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface UpdateCredentialsInput {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 3 */
  newUsername?: string;
  /** @minLength 6 */
  newPassword?: string;
}

export interface Settings {
  businessName: string;
  /** @nullable */
  logoUrl?: string | null;
}

export interface UpdateSettingsInput {
  /** @minLength 1 */
  businessName?: string;
  /** @nullable */
  logoUrl?: string | null;
}

export interface UploadUrlRequest {
  /**
   * Original file name.
   * @minLength 1
   */
  name: string;
  /**
   * File size in bytes.
   * @minimum 1
   */
  size: number;
  /**
   * MIME type of the file (e.g. `image/jpeg`).
   * @minLength 1
   */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  sku: string | null;
  /** @nullable */
  description: string | null;
  /** Price in KES cents */
  priceCents: number;
  stock: number;
  /** @nullable */
  category: string | null;
  /** @nullable */
  imageUrl: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  priceCents: number;
  /** @minimum 0 */
  stock: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface UpdateProductInput {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  priceCents?: number;
  /** @minimum 0 */
  stock?: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPriceCents: number;
  lineTotalCents: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  paid: "paid",
  cancelled: "cancelled",
  failed: "failed",
} as const;

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  mpesa: "mpesa",
  cash: "cash",
} as const;

export interface Order {
  id: number;
  reference: string;
  status: OrderStatus;
  totalCents: number;
  /** @nullable */
  customerPhone: string | null;
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  cashierName: string | null;
  paymentMethod: OrderPaymentMethod;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export type CreateOrderInputPaymentMethod =
  (typeof CreateOrderInputPaymentMethod)[keyof typeof CreateOrderInputPaymentMethod];

export const CreateOrderInputPaymentMethod = {
  mpesa: "mpesa",
  cash: "cash",
} as const;

export interface CreateOrderInput {
  /** @minItems 1 */
  items: CreateOrderItemInput[];
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  cashierName?: string | null;
  paymentMethod: CreateOrderInputPaymentMethod;
}

export type TransactionProvider =
  (typeof TransactionProvider)[keyof typeof TransactionProvider];

export const TransactionProvider = {
  mpesa: "mpesa",
  cash: "cash",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  success: "success",
  failed: "failed",
} as const;

export interface Transaction {
  id: number;
  /** @nullable */
  orderId: number | null;
  /** @nullable */
  orderReference: string | null;
  amountCents: number;
  provider: TransactionProvider;
  status: TransactionStatus;
  /** @nullable */
  mpesaCheckoutRequestId: string | null;
  /** @nullable */
  mpesaReceiptNumber: string | null;
  /** @nullable */
  phoneNumber: string | null;
  /** @nullable */
  rawCallback: string | null;
  createdAt: string;
}

export interface StkPushInput {
  orderId: number;
  /** MSISDN in format 2547XXXXXXXX */
  phoneNumber: string;
}

export interface StkPushResponse {
  /** @nullable */
  merchantRequestId: string | null;
  /** @nullable */
  checkoutRequestId: string | null;
  /** @nullable */
  responseCode: string | null;
  /** @nullable */
  responseDescription: string | null;
  /** @nullable */
  customerMessage: string | null;
  transactionId: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  unitsSold: number;
  revenueCents: number;
}

export interface SalesByDay {
  /** ISO date (YYYY-MM-DD) */
  day: string;
  revenueCents: number;
  orderCount: number;
}

export interface DashboardSummary {
  totalRevenueCents: number;
  todayRevenueCents: number;
  totalOrders: number;
  paidOrders: number;
  pendingOrders: number;
  totalProducts: number;
  lowStockCount: number;
  recentOrders: Order[];
  topProducts: TopProduct[];
  salesByDay: SalesByDay[];
}

export type ListProductsParams = {
  search?: string;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  paid: "paid",
  cancelled: "cancelled",
  failed: "failed",
} as const;