# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Duka POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mpesa/stkpush:
    post:
      operationId: mpesaStkPush
      tags: [mpesa]
      summary: Initiate M-Pesa STK Push
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StkPushInput"
      responses:
        "200":
          description: STK Push initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StkPushResponse"
        "400":
          description: Invalid input or M-Pesa not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Daraja API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mpesa/callback:
    post:
      operationId: mpesaCallback
      tags: [mpesa]
      summary: M-Pesa STK Push callback (called by Safaricom Daraja)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MpesaCallbackPayload"
      responses:
        "200":
          description: Callback acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpesaCallbackAck"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/credentials:
    patch:
      operationId: updateCredentials
      tags: [auth]
      summary: Change username and/or password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCredentialsInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated or current password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    MpesaCallbackPayload:
      type: object
      properties:
        Body:
          type: object
          additionalProperties: true
      required:
        - Body
      additionalProperties: true
    MpesaCallbackAck:
      type: object
      properties:
        ResultCode:
          type: integer
        ResultDesc:
          type: string
      required:
        - ResultCode
        - ResultDesc
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    UpdateCredentialsInput:
      type: object
      required: [currentPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newUsername:
          type: string
          minLength: 3
        newPassword:
          type: string
          minLength: 6
...
    StkPushInput:
      type: object
      properties:
        orderId:
          type: integer
        phoneNumber:
          type: string
          description: MSISDN in format 2547XXXXXXXX
      required:
        - orderId
        - phoneNumber
    StkPushResponse:
      type: object
      properties:
        merchantRequestId:
          type: ["string", "null"]
        checkoutRequestId:
          type: ["string", "null"]
        responseCode:
          type: ["string", "null"]
        responseDescription:
          type: ["string", "null"]
        customerMessage:
          type: ["string", "null"]
        transactionId:
          type: integer
      required:
        - merchantRequestId
        - checkoutRequestId
        - responseCode
        - responseDescription
        - customerMessage
        - transactionId
...