# lib/api-zod/src/generated/types/mpesaCallbackPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Duka POS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MpesaCallbackPayloadBody } from "./mpesaCallbackPayloadBody";

export interface MpesaCallbackPayload {
  Body: MpesaCallbackPayloadBody;
  [key: string]: unknown;
}