# lib/api-zod/src/generated/types/stkPushResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Duka POS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StkPushResponse {
  /** @nullable */
  merchantRequestId: string | null;
  /** @nullable */
  checkoutRequestId: string | null;
  /** @nullable */
  responseCode: string | null;
  /** @nullable */
  responseDescription: string | null;
  /** @nullable */
  customerMessage: string | null;
  transactionId: number;
}